Add yarn.lock
This is from the production server and we need to use it to install
the exact versions of deps that are working rather than evaluating
all the semver strings for all deps and hitting issues.

Install exact deps with yarn using this:

    yarn install --frozen-lockfile

This is similar to `npm ci` for installing the exact deps from the
package-lock.json.